requirements.txt
streamlit
google-generativeai
pymupdf
gspread
google-auth